+++ mobilitydb/test/general/queries/038_temporal_similarity_tbl.test.sql
@@ -31,8 +31,8 @@
 -- Discrete Frechet distance and path
 -------------------------------------------------------------------------------
 
-SELECT round(MAX(frechetDistance(t1.temp, t2.temp))::numeric, 6) FROM tbl_tint t1, tbl_tint t2 WHERE t1.k < t2.k;
-SELECT round(MAX(frechetDistance(t1.temp, t2.temp))::numeric, 6) FROM tbl_tfloat t1, tbl_tfloat t2 WHERE t1.k < t2.k;
+SELECT round(MAX(frechetDistance(t1.temp, t2.temp)), 6) FROM tbl_tint t1, tbl_tint t2 WHERE t1.k < t2.k;
+SELECT round(MAX(frechetDistance(t1.temp, t2.temp)), 6) FROM tbl_tfloat t1, tbl_tfloat t2 WHERE t1.k < t2.k;
 
 WITH temp AS (
   SELECT frechetDistancePath(t1.temp, t2.temp) FROM tbl_tint t1, tbl_tint t2 WHERE t1.k < t2.k )
@@ -45,8 +45,8 @@ SELECT COUNT(*) FROM temp;
 -- Dynamic Time Warp (DTW) distance and path
 -------------------------------------------------------------------------------
 
-SELECT round(MAX(dynTimeWarpDistance(t1.temp, t2.temp))::numeric, 6) FROM tbl_tint t1, tbl_tint t2 WHERE t1.k < t2.k;
-SELECT round(MAX(dynTimeWarpDistance(t1.temp, t2.temp))::numeric, 6) FROM tbl_tfloat t1, tbl_tfloat t2 WHERE t1.k < t2.k;
+SELECT round(MAX(dynTimeWarpDistance(t1.temp, t2.temp)), 6) FROM tbl_tint t1, tbl_tint t2 WHERE t1.k < t2.k;
+SELECT round(MAX(dynTimeWarpDistance(t1.temp, t2.temp)), 6) FROM tbl_tfloat t1, tbl_tfloat t2 WHERE t1.k < t2.k;
 
 WITH temp AS (
   SELECT dynTimeWarpPath(t1.temp, t2.temp) FROM tbl_tint t1, tbl_tint t2 WHERE t1.k < t2.k )
@@ -59,7 +59,7 @@ SELECT COUNT(*) FROM temp;
 -- Hausdorff distance
 -------------------------------------------------------------------------------
 
-SELECT round(MAX(hausdorffDistance(t1.temp, t2.temp))::numeric, 6) FROM tbl_tint t1, tbl_tint t2 WHERE t1.k < t2.k;
-SELECT round(MAX(hausdorffDistance(t1.temp, t2.temp))::numeric, 6) FROM tbl_tfloat t1, tbl_tfloat t2 WHERE t1.k < t2.k;
+SELECT round(MAX(hausdorffDistance(t1.temp, t2.temp)), 6) FROM tbl_tint t1, tbl_tint t2 WHERE t1.k < t2.k;
+SELECT round(MAX(hausdorffDistance(t1.temp, t2.temp)), 6) FROM tbl_tfloat t1, tbl_tfloat t2 WHERE t1.k < t2.k;
 
 -------------------------------------------------------------------------------
